+++ Disasteroids/trunk/src/disasteroids/LaserManager.java
@@ -17,7 +17,6 @@ import java.util.concurrent.ConcurrentLinkedQueue;
  */
 class LaserManager extends Weapon
 {
-
     private int speed = 20;
 
     private int maxShots = 1000;
@@ -25,7 +24,7 @@ class LaserManager extends Weapon
     private int intervalShoot = 10;
 
     private int damage = 10;
-    
+
     private int length = 10;
 
     public LaserManager()
@@ -40,10 +39,9 @@ class LaserManager extends Weapon
 
     public void act()
     {
-        super.act(true);
+        super.act( true );
     }
 
-  
     public int getIntervalShoot()
     {
         return intervalShoot;
@@ -54,19 +52,19 @@ class LaserManager extends Weapon
         if ( weapons.size() > 2500 || timeTillNextShot > 0 )
             return false;
         timeTillNextShot = intervalShoot;
-        
-        int X=x;
-        int Y=y;
-        Laser l=null;
-        for(int i=0; i<150; i++)
-        { 
-            Laser last=new Laser(this, X, Y, angle, dx, dy, col);
-            if(l!=null)
-                l.setNext(last);
-            weapons.add(last);
-            X+=length*Math.cos(angle);
-            Y-=length*Math.sin(angle);
-            l=last;
+
+        int X = x;
+        int Y = y;
+        Laser l = null;
+        for ( int i = 0; i < 150; i++ )
+        {
+            Laser last = new Laser( this, X, Y, angle, dx, dy, col );
+            if ( l != null )
+                l.setNext( last );
+            weapons.add( last );
+            X += length * Math.cos( angle );
+            Y -= length * Math.sin( angle );
+            l = last;
         }
         if ( playShootSound )
             Sound.playInternal( getShootSound() );
@@ -76,7 +74,7 @@ class LaserManager extends Weapon
 
     public void restoreBonusValues()
     {
-        
+
     }
 
     public int getDamage()
@@ -108,7 +106,7 @@ class LaserManager extends Weapon
 
     public int getRadius()
     {
-        return length/2;
+        return length / 2;
     }
 
     public void draw( Graphics g )
@@ -136,13 +134,13 @@ class LaserManager extends Weapon
         timeTillNextShot = 0;
         for ( double angle = 0; angle < 2 * Math.PI; angle += Math.PI / 5 )
         {
-            add( s.getX(), s.getY(), angle, s.getDx(), s.getDy(), s.getColor(), false );
+            add( (int) s.getX(), (int) s.getY(), angle, s.getDx(), s.getDy(), s.getColor(), false );
             timeTillNextShot = 0;
         }
         timeTillNextShot = temp;
         timeTillNextBerserk = 50;
     }
-    
+
     public int length()
     {
         return length;
@@ -151,9 +149,8 @@ class LaserManager extends Weapon
     @Override
     public boolean canShoot()
     {
-        return super.canShoot()&& weapons.size() < 500 ;
+        return super.canShoot() && weapons.size() < 500;
     }
-
 
     public SoundClip getShootSound()
     {
